ecs: Resolve Clippy lints
There was one false positive never_loop lint.

+++ quill/ecs/tests/ecs.rs
@@ -23,14 +23,14 @@ fn spawn_many_entities() {
     let mut entities = Vec::new();
     for i in 0..10_000 {
         let entity = EntityBuilder::new()
-            .add(10.0f32)
+            .add(10i32)
             .add(format!("Entity #{}", i))
             .spawn_into(&mut ecs);
         entities.push(entity);
     }
 
     for (i, entity) in entities.into_iter().enumerate() {
-        assert_eq!(*ecs.get::<f32>(entity).unwrap(), 10.0);
+        assert_eq!(*ecs.get::<i32>(entity).unwrap(), 10);
         assert_eq!(
             *ecs.get::<String>(entity).unwrap(),
             format!("Entity #{}", i)
